Convert status stop
